+++ projects/igniteui-angular/src/lib/grid/grid.groupby.spec.ts
@@ -129,7 +129,7 @@ describe('IgxGrid - GroupBy', () => {
 
         // group by string column
         const grid = fix.componentInstance.instance;
-        grid.groupBy('ProductName', SortingDirection.Desc, false);
+        grid.groupBy({fieldName: 'ProductName', dir: SortingDirection.Desc, ignoreCase: false});
         fix.detectChanges();
 
         // verify grouping expressions
@@ -154,7 +154,7 @@ describe('IgxGrid - GroupBy', () => {
         expect(grid.groupedRowList.toArray().length).toEqual(0);
 
         // group by number
-        grid.groupBy('Downloads', SortingDirection.Desc, false);
+        grid.groupBy({fieldName: 'Downloads', dir: SortingDirection.Desc, ignoreCase: false});
         fix.detectChanges();
 
         groupRows = grid.groupedRowList.toArray();
@@ -168,7 +168,7 @@ describe('IgxGrid - GroupBy', () => {
         // ungroup and group by boolean column
         grid.clearGrouping('Downloads');
         fix.detectChanges();
-        grid.groupBy('Released', SortingDirection.Desc, false);
+        grid.groupBy({fieldName: 'Released', dir: SortingDirection.Desc, ignoreCase: false});
         fix.detectChanges();
 
         groupRows = grid.groupedRowList.toArray();
@@ -182,7 +182,7 @@ describe('IgxGrid - GroupBy', () => {
         // ungroup and group by date column
         grid.clearGrouping('Released');
         fix.detectChanges();
-        grid.groupBy('ReleaseDate', SortingDirection.Asc, false);
+        grid.groupBy({fieldName: 'ReleaseDate', dir: SortingDirection.Asc, ignoreCase: false});
         fix.detectChanges();
 
         groupRows = grid.groupedRowList.toArray();
@@ -203,8 +203,8 @@ describe('IgxGrid - GroupBy', () => {
 
         // group by 2 columns
         const grid = fix.componentInstance.instance;
-        grid.groupBy('ProductName', SortingDirection.Desc, false);
-        grid.groupBy('Released', SortingDirection.Desc, false);
+        grid.groupBy({fieldName: 'ProductName', dir: SortingDirection.Desc, ignoreCase: false});
+        grid.groupBy({fieldName: 'Released', dir: SortingDirection.Desc, ignoreCase: false});
         fix.detectChanges();
 
         let groupRows = grid.groupedRowList.toArray();
@@ -221,7 +221,7 @@ describe('IgxGrid - GroupBy', () => {
 
         // group by 3rd column
 
-        grid.groupBy('Downloads', SortingDirection.Desc, false);
+        grid.groupBy({fieldName: 'Downloads', dir: SortingDirection.Desc, ignoreCase: false});
         fix.detectChanges();
 
         groupRows = grid.groupedRowList.toArray();
@@ -242,7 +242,7 @@ describe('IgxGrid - GroupBy', () => {
         const grid = fix.componentInstance.instance;
         grid.primaryKey = 'ID';
         fix.detectChanges();
-        grid.groupBy('Released', SortingDirection.Desc, false);
+        grid.groupBy({fieldName: 'Released', dir: SortingDirection.Desc, ignoreCase: false});
         fix.detectChanges();
 
         let groupRows = grid.groupedRowList.toArray();
@@ -335,7 +335,7 @@ describe('IgxGrid - GroupBy', () => {
         fix.detectChanges();
 
         grid.groupsExpanded = false;
-        grid.groupBy('Released', SortingDirection.Desc, false);
+        grid.groupBy({ fieldName: 'Released', dir: SortingDirection.Desc, ignoreCase: false });
         fix.detectChanges();
 
         let groupRows = grid.groupedRowList.toArray();
@@ -368,7 +368,7 @@ describe('IgxGrid - GroupBy', () => {
         grid.primaryKey = 'ID';
         fix.detectChanges();
 
-        grid.groupBy('Released', SortingDirection.Desc, false);
+        grid.groupBy({fieldName: 'Released', dir: SortingDirection.Desc, ignoreCase: false});
         fix.detectChanges();
 
         const currExpr = fix.componentInstance.currentSortExpressions;
@@ -381,7 +381,7 @@ describe('IgxGrid - GroupBy', () => {
         const grid = fix.componentInstance.instance;
         fix.detectChanges();
 
-        grid.groupBy('Released', SortingDirection.Desc, false);
+        grid.groupBy({fieldName: 'Released', dir: SortingDirection.Desc, ignoreCase: false});
         fix.detectChanges();
 
         const groupRows = grid.groupedRowList.toArray();
@@ -401,7 +401,7 @@ describe('IgxGrid - GroupBy', () => {
         grid.primaryKey = 'ID';
         fix.detectChanges();
 
-        grid.groupBy('Released', SortingDirection.Desc, false);
+        grid.groupBy({fieldName: 'Released', dir: SortingDirection.Desc, ignoreCase: false});
         fix.detectChanges();
 
         const groupRows = grid.groupedRowList.toArray();
@@ -418,7 +418,7 @@ describe('IgxGrid - GroupBy', () => {
         const grid = fix.componentInstance.instance;
         fix.componentInstance.enableSorting = true;
         fix.detectChanges();
-        grid.groupBy('ProductName', SortingDirection.Desc, false);
+        grid.groupBy({fieldName: 'ProductName', dir: SortingDirection.Desc, ignoreCase: false});
         fix.detectChanges();
         const groupRows = grid.groupedRowList.toArray();
         const dataRows = grid.dataRowList.toArray();
@@ -426,7 +426,7 @@ describe('IgxGrid - GroupBy', () => {
         expect(groupRows.length).toEqual(5);
         expect(dataRows.length).toEqual(8);
 
-        grid.sort('Released', SortingDirection.Asc, false);
+        grid.sort({fieldName: 'Released', dir: SortingDirection.Asc, ignoreCase: false});
         fix.detectChanges();
 
         // verify groups
@@ -445,7 +445,7 @@ describe('IgxGrid - GroupBy', () => {
         const grid = fix.componentInstance.instance;
         fix.componentInstance.enableSorting = true;
         fix.detectChanges();
-        grid.groupBy('ProductName', SortingDirection.Desc, false);
+        grid.groupBy({fieldName: 'ProductName', dir: SortingDirection.Desc, ignoreCase: false});
         fix.detectChanges();
 
         let groupRows = grid.groupedRowList.toArray();
@@ -457,7 +457,7 @@ describe('IgxGrid - GroupBy', () => {
 
         // verify group order
         checkGroups(groupRows, ['NetAdvantage', 'Ignite UI for JavaScript', 'Ignite UI for Angular', '', null]);
-        grid.sort('ProductName', SortingDirection.Asc, false);
+        grid.sort({fieldName: 'ProductName', dir: SortingDirection.Asc, ignoreCase: false});
         fix.detectChanges();
 
         groupRows = grid.groupedRowList.toArray();
@@ -473,7 +473,7 @@ describe('IgxGrid - GroupBy', () => {
         const grid = fix.componentInstance.instance;
         fix.componentInstance.enableSorting = true;
         fix.detectChanges();
-        grid.groupBy('ProductName', SortingDirection.Desc, false);
+        grid.groupBy({fieldName: 'ProductName', dir: SortingDirection.Desc, ignoreCase: false});
         fix.detectChanges();
 
         let groupRows = grid.groupedRowList.toArray();
@@ -485,7 +485,7 @@ describe('IgxGrid - GroupBy', () => {
 
         // verify group order
         checkGroups(groupRows, ['NetAdvantage', 'Ignite UI for JavaScript', 'Ignite UI for Angular', '', null]);
-        grid.sort('ProductName', SortingDirection.None, false);
+        grid.sort({fieldName: 'ProductName', dir: SortingDirection.None, ignoreCase: false});
         fix.detectChanges();
         groupRows = grid.groupedRowList.toArray();
         dataRows = grid.dataRowList.toArray();
@@ -501,7 +501,7 @@ describe('IgxGrid - GroupBy', () => {
         const grid = fix.componentInstance.instance;
         fix.componentInstance.enableSorting = true;
         fix.detectChanges();
-        grid.groupBy('Downloads', SortingDirection.Desc, false);
+        grid.groupBy({fieldName: 'Downloads', dir: SortingDirection.Desc, ignoreCase: false});
         fix.detectChanges();
 
         const headers = fix.debugElement.queryAll(By.css(COLUMN_HEADER_CLASS));
@@ -530,10 +530,10 @@ describe('IgxGrid - GroupBy', () => {
         const grid = fix.componentInstance.instance;
         fix.componentInstance.enableSorting = true;
         fix.detectChanges();
-        grid.sort('ProductName', SortingDirection.Desc, false);
+        grid.sort({fieldName: 'ProductName', dir: SortingDirection.Desc, ignoreCase: false});
         fix.detectChanges();
 
-        grid.groupBy('ProductName', SortingDirection.Asc, false);
+        grid.groupBy({fieldName: 'ProductName', dir: SortingDirection.Asc, ignoreCase: false});
         fix.detectChanges();
         const groupRows = grid.groupedRowList.toArray();
         // verify group order
@@ -545,9 +545,9 @@ describe('IgxGrid - GroupBy', () => {
         const grid = fix.componentInstance.instance;
         fix.componentInstance.enableSorting = true;
         fix.detectChanges();
-        grid.sort('ProductName', SortingDirection.Desc, false);
+        grid.sort({ fieldName: 'ProductName', dir: SortingDirection.Desc, ignoreCase: false });
         fix.detectChanges();
-        grid.groupBy('ProductName', SortingDirection.Desc, false);
+        grid.groupBy({ fieldName: 'ProductName', dir: SortingDirection.Desc, ignoreCase: false});
         fix.detectChanges();
         const groupRows = grid.groupedRowList.toArray();
         const dataRows = grid.dataRowList.toArray();
@@ -564,7 +564,7 @@ describe('IgxGrid - GroupBy', () => {
         fix.componentInstance.width = '400px';
         fix.detectChanges();
 
-        grid.groupBy('ProductName', SortingDirection.Desc, false);
+        grid.groupBy({fieldName: 'ProductName', dir: SortingDirection.Desc, ignoreCase: false});
         fix.detectChanges();
         const gRow = grid.groupedRowList.toArray()[0];
         expect(gRow.expanded).toBe(true);
@@ -597,8 +597,8 @@ describe('IgxGrid - GroupBy', () => {
         grid.columnWidth = '200px';
         fix.detectChanges();
 
-        grid.groupBy('ProductName', SortingDirection.Desc, false);
-        grid.groupBy('Released', SortingDirection.Desc, false);
+        grid.groupBy({fieldName: 'ProductName', dir: SortingDirection.Desc, ignoreCase: false});
+        grid.groupBy({fieldName: 'Released', dir: SortingDirection.Desc, ignoreCase: false});
         fix.detectChanges();
         const cbFunc2 = () => {
             const row = grid.getRowByIndex(0);
@@ -629,8 +629,8 @@ describe('IgxGrid - GroupBy', () => {
         grid.columnWidth = '200px';
         fix.detectChanges();
 
-        grid.groupBy('ProductName', SortingDirection.Desc, false);
-        grid.groupBy('Released', SortingDirection.Desc, false);
+        grid.groupBy({fieldName: 'ProductName', dir: SortingDirection.Desc, ignoreCase: false});
+        grid.groupBy({fieldName: 'Released', dir: SortingDirection.Desc, ignoreCase: false});
         fix.detectChanges();
         grid.parentVirtDir.getHorizontalScroll().scrollLeft = 1000;
         fix.detectChanges();
@@ -668,8 +668,8 @@ describe('IgxGrid - GroupBy', () => {
         grid.columnWidth = '200px';
         fix.detectChanges();
 
-        grid.groupBy('ProductName', SortingDirection.Desc, false);
-        grid.groupBy('Released', SortingDirection.Desc, false);
+        grid.groupBy({fieldName: 'ProductName', dir: SortingDirection.Desc, ignoreCase: false});
+        grid.groupBy({fieldName: 'Released', dir: SortingDirection.Desc, ignoreCase: false});
         fix.detectChanges();
 
         const cbFunc = () => {
@@ -700,8 +700,8 @@ describe('IgxGrid - GroupBy', () => {
         grid.columnWidth = '200px';
         fix.detectChanges();
 
-        grid.groupBy('ProductName', SortingDirection.Desc, false);
-        grid.groupBy('Released', SortingDirection.Desc, false);
+        grid.groupBy({fieldName: 'ProductName', dir: SortingDirection.Desc, ignoreCase: false});
+        grid.groupBy({fieldName: 'Released', dir: SortingDirection.Desc, ignoreCase: false});
         fix.detectChanges();
 
         expect(grid.groupedRowList.toArray().length).toEqual(3);
@@ -718,8 +718,8 @@ describe('IgxGrid - GroupBy', () => {
         grid.columnWidth = '200px';
         fix.detectChanges();
 
-        grid.groupBy('ProductName', SortingDirection.Desc, false);
-        grid.groupBy('Released', SortingDirection.Desc, false);
+        grid.groupBy({fieldName: 'ProductName', dir: SortingDirection.Desc, ignoreCase: false});
+        grid.groupBy({fieldName: 'Released', dir: SortingDirection.Desc, ignoreCase: false});
         fix.detectChanges();
 
         const origScrollHeight = parseInt(grid.verticalScrollContainer.getVerticalScroll().children[0].style.height, 10);
@@ -758,8 +758,8 @@ describe('IgxGrid - GroupBy', () => {
         grid.columnWidth = '200px';
         fix.detectChanges();
 
-        grid.groupBy('ProductName', SortingDirection.Desc, false);
-        grid.groupBy('Released', SortingDirection.Desc, false);
+        grid.groupBy({ fieldName: 'ProductName', dir: SortingDirection.Desc, ignoreCase: false});
+        grid.groupBy({ fieldName: 'Released', dir: SortingDirection.Desc, ignoreCase: false});
         fix.detectChanges();
 
         let groupRow = grid.groupedRowList.toArray()[0];
@@ -790,8 +790,8 @@ describe('IgxGrid - GroupBy', () => {
         grid.columnWidth = '200px';
         fix.detectChanges();
 
-        grid.groupBy('ProductName', SortingDirection.Desc, false);
-        grid.groupBy('Released', SortingDirection.Desc, false);
+        grid.groupBy({fieldName: 'ProductName', dir: SortingDirection.Desc, ignoreCase: false});
+        grid.groupBy({fieldName: 'Released', dir: SortingDirection.Desc, ignoreCase: false});
         fix.detectChanges();
         const groupRow = grid.groupedRowList.toArray()[0];
         const origRect = groupRow.element.nativeElement.getBoundingClientRect();
@@ -813,7 +813,7 @@ describe('IgxGrid - GroupBy', () => {
         grid.columnWidth = '200px';
         fix.detectChanges();
 
-        grid.groupBy('ProductName', SortingDirection.Desc, false);
+        grid.groupBy({fieldName: 'ProductName', dir: SortingDirection.Desc, ignoreCase: false});
         let groupRows = grid.groupedRowList.toArray();
         let dataRows = grid.dataRowList.toArray();
         let allRows = grid.rowList.toArray();
@@ -844,7 +844,7 @@ describe('IgxGrid - GroupBy', () => {
         grid.rowSelectable = true;
         fix.detectChanges();
 
-        grid.groupBy('ProductName', SortingDirection.Desc, false);
+        grid.groupBy({fieldName: 'ProductName',  dir: SortingDirection.Desc, ignoreCase: false});
 
         fix.detectChanges();
 
@@ -868,7 +868,7 @@ describe('IgxGrid - GroupBy', () => {
         grid.rowSelectable = true;
         fix.detectChanges();
 
-        grid.groupBy('ProductName', SortingDirection.Desc, false);
+        grid.groupBy({fieldName: 'ProductName', dir: SortingDirection.Desc, ignoreCase: false});
 
         fix.detectChanges();
 
@@ -912,7 +912,7 @@ describe('IgxGrid - GroupBy', () => {
         fix.componentInstance.enableResizing = true;
         grid.columnWidth = '200px';
         fix.detectChanges();
-        grid.groupBy('ProductName', SortingDirection.Desc, false);
+        grid.groupBy({fieldName: 'ProductName', dir: SortingDirection.Desc, ignoreCase: false});
         fix.detectChanges();
 
         let grRows = grid.groupedRowList.toArray();
@@ -956,7 +956,7 @@ describe('IgxGrid - GroupBy', () => {
         fix.componentInstance.width = '1200px';
         grid.columnWidth = '200px';
         fix.detectChanges();
-        grid.groupBy('ProductName', SortingDirection.Desc, false);
+        grid.groupBy({fieldName: 'ProductName', dir: SortingDirection.Desc, ignoreCase: false});
         fix.detectChanges();
 
         grid.enableSummaries([{ fieldName: 'ProductName' }]);
@@ -980,7 +980,7 @@ describe('IgxGrid - GroupBy', () => {
         fix.componentInstance.width = '1200px';
         grid.columnWidth = '200px';
         fix.detectChanges();
-        grid.groupBy('ProductName', SortingDirection.Desc, false);
+        grid.groupBy({fieldName: 'ProductName', dir: SortingDirection.Desc, ignoreCase: false});
         fix.detectChanges();
 
         grid.getColumnByName('ProductName').hidden = true;
@@ -1001,7 +1001,7 @@ describe('IgxGrid - GroupBy', () => {
         fix.componentInstance.width = '500px';
         grid.columnWidth = '200px';
         fix.detectChanges();
-        grid.groupBy('ProductName', SortingDirection.Desc, false);
+        grid.groupBy({fieldName: 'ProductName', dir: SortingDirection.Desc, ignoreCase: false});
         fix.detectChanges();
 
         grid.pinColumn('ProductName');
@@ -1021,7 +1021,7 @@ describe('IgxGrid - GroupBy', () => {
         grid.columnWidth = '200px';
         grid.primaryKey = 'ID';
         fix.detectChanges();
-        grid.groupBy('ProductName', SortingDirection.Desc, false);
+        grid.groupBy({fieldName: 'ProductName', dir: SortingDirection.Desc, ignoreCase: false});
         fix.detectChanges();
 
         // verify rows
@@ -1072,7 +1072,7 @@ describe('IgxGrid - GroupBy', () => {
         grid.columnWidth = '200px';
         grid.primaryKey = 'ID';
         fix.detectChanges();
-        grid.groupBy('ProductName', SortingDirection.Desc, false);
+        grid.groupBy({fieldName: 'ProductName', dir: SortingDirection.Desc, ignoreCase: false});
         fix.detectChanges();
 
         const rv = grid.getRowByKey(5).element.nativeElement.querySelectorAll(CELL_CSS_CLASS)[2];
@@ -1103,7 +1103,7 @@ describe('IgxGrid - GroupBy', () => {
         expect(dataRows.length).toEqual(8);
 
         // re-apply grouping
-        grid.groupBy('ProductName', SortingDirection.Asc, false);
+        grid.groupBy({fieldName: 'ProductName', dir: SortingDirection.Asc, ignoreCase: false});
         fix.detectChanges();
 
         groupRows = grid.groupedRowList.toArray();
@@ -1119,7 +1119,7 @@ describe('IgxGrid - GroupBy', () => {
         fix.componentInstance.instance.paging = true;
         fix.componentInstance.instance.perPage = 3;
         fix.detectChanges();
-        fix.componentInstance.instance.groupBy('ProductName', SortingDirection.Desc, false);
+        fix.componentInstance.instance.groupBy({fieldName: 'ProductName', dir: SortingDirection.Desc, ignoreCase: false});
         const groupRows = grid.groupedRowList.toArray();
         const dataRows = grid.dataRowList.toArray();
 
@@ -1136,7 +1136,7 @@ describe('IgxGrid - GroupBy', () => {
         fix.componentInstance.instance.paging = true;
         fix.componentInstance.instance.perPage = 3;
         fix.detectChanges();
-        fix.componentInstance.instance.groupBy('ProductName', SortingDirection.Desc, false);
+        fix.componentInstance.instance.groupBy({fieldName: 'ProductName', dir: SortingDirection.Desc, ignoreCase: false});
         let groupRows = grid.groupedRowList.toArray();
         let dataRows = grid.dataRowList.toArray();
 
@@ -1171,7 +1171,7 @@ describe('IgxGrid - GroupBy', () => {
             value: 'Ignite UI for JavaScript'
         });
         fix.detectChanges();
-        fix.componentInstance.instance.groupBy('ProductName', SortingDirection.Desc, false);
+        fix.componentInstance.instance.groupBy({fieldName: 'ProductName', dir: SortingDirection.Desc, ignoreCase: false});
         let groupRows = grid.groupedRowList.toArray();
         let dataRows = grid.dataRowList.toArray();
 
@@ -1212,7 +1212,7 @@ describe('IgxGrid - GroupBy', () => {
         fix.detectChanges();
         const gridElement: HTMLElement = fix.nativeElement.querySelector('.igx-grid');
 
-        grid.groupBy('ProductName', SortingDirection.Asc, false);
+        grid.groupBy({fieldName: 'ProductName', dir: SortingDirection.Asc, ignoreCase: false});
         fix.detectChanges();
         const groupRows = grid.groupedRowList.toArray();
         // verify group area is rendered
@@ -1235,7 +1235,7 @@ describe('IgxGrid - GroupBy', () => {
         fix.detectChanges();
         const gridElement: HTMLElement = fix.nativeElement.querySelector('.igx-grid');
 
-        grid.groupBy('ProductName', SortingDirection.Asc, false);
+        grid.groupBy({fieldName: 'ProductName', dir: SortingDirection.Asc, ignoreCase: false});
         grid.toggleAllGroupRows();
         fix.detectChanges();
         const groupRows = grid.groupedRowList.toArray();